tests and st working together fix
