+++ tests/unit/modules/test_state.py
@@ -1,4 +1,4 @@
-from code.modules.state import (
+from modules.state import (
     read_environ_params,
     read_url_param_values
     )
